pyproject.toml
[project]
name = "pml-i2sb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "diffusers>=0.35.2",
    "pillow>=12.0.0",
    "pyyaml>=6.0.3",
    "tensorflow>=2.20.0",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
]
